+++ CHANGES.md
@@ -2,11 +2,13 @@
 
 ## 0.12.0 (2022-02-XX)
 
+- **ENH: Adding the support of `Vision-1` sensor**
 - **ENH: Adding a keyword to allow the pass of a specific DEM path in `load`/`stack` functions (for VHR orthorectification and `DEM` bands)**
 - FIX: Using `COPDEM-30` (`GLO-30`) by default for SNAP as it appears that the retrieval has been fixed.
 - FIX: Fixing the default name for cleaned bands for `Sentinel-3 SLSTR` data (was set on `CLEAN` instead of `NODATA`)
 - FIX: Adding the name of the DEM in DEM band (i.e. allow to compute the `HILLSHADE` with a DEM and the `SLOPE` with a DTM)
 - FIX: `Sentinel-2` Processing Baseline 04.00: `NARROW_NIR` bands are now loaded correctly
+- FIX: `Maxar` products (with `Multi` band ID) are now correctly handled
 - OPTIM: Do not look for valid metadata further than a given nested level in product's directory (for extracted products)
 - CI: Using another runner
 - DOC: Adding a DEM notebook